# docs/contracts/backtest-service.openapi.json
{
  "components": {
    "schemas": {
      "BacktestExecuteRequest": {
        "additionalProperties": false,
        "description": "Bounded deterministic calculation request from the control plane/Agent.",
        "properties": {
          "deadline_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline At"
          },
          "expected_data_version": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 3,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Data Version"
          },
          "idempotency_key": {
            "maxLength": 128,
            "minLength": 16,
            "title": "Idempotency Key",
            "type": "string"
          },
          "job_id": {
            "maxLength": 128,
            "minLength": 8,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "Job Id",
            "type": "string"
          },
          "max_bars": {
            "default": 250000,
            "maximum": 500000.0,
            "minimum": 100.0,
            "title": "Max Bars",
            "type": "integer"
          },
          "max_trades": {
            "default": 5000,
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Max Trades",
            "type": "integer"
          },
          "request_id": {
            "maxLength": 128,
            "minLength": 8,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "Request Id",
            "type": "string"
          },
          "requested_at": {
            "format": "date-time",
            "title": "Requested At",
            "type": "string"
          },
          "run_id": {
            "maxLength": 128,
            "minLength": 8,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "Run Id",
            "type": "string"
          },
          "schema_version": {
            "const": "backtest-request-v1",
            "default": "backtest-request-v1",
            "title": "Schema Version",
            "type": "string"
          },
          "strategy": {
            "$ref": "#/components/schemas/StrategySpec"
          }
        },
        "required": [
          "request_id",
          "job_id",
          "run_id",
          "idempotency_key",
          "strategy"
        ],
        "title": "BacktestExecuteRequest",
        "type": "object"
      },
      "BacktestExecuteResponse": {
        "additionalProperties": false,
        "properties": {
          "data_profile": {
            "$ref": "#/components/schemas/DataProfile"
          },
          "data_version": {
            "maxLength": 128,
            "minLength": 3,
            "title": "Data Version",
            "type": "string"
          },
          "duration_ms": {
            "minimum": 0.0,
            "title": "Duration Ms",
            "type": "number"
          },
          "engine_version": {
            "const": "sma-crossover-v1",
            "default": "sma-crossover-v1",
            "title": "Engine Version",
            "type": "string"
          },
          "equity_curve": {
            "items": {
              "$ref": "#/components/schemas/EquityPoint"
            },
            "maxItems": 321,
            "title": "Equity Curve",
            "type": "array"
          },
          "job_id": {
            "maxLength": 128,
            "minLength": 8,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "Job Id",
            "type": "string"
          },
          "metrics": {
            "$ref": "#/components/schemas/BacktestMetrics"
          },
          "request_id": {
            "maxLength": 128,
            "minLength": 8,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "Request Id",
            "type": "string"
          },
          "result_digest": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Result Digest",
            "type": "string"
          },
          "run_id": {
            "maxLength": 128,
            "minLength": 8,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "Run Id",
            "type": "string"
          },
          "schema_version": {
            "const": "backtest-result-v1",
            "default": "backtest-result-v1",
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "const": "completed",
            "default": "completed",
            "title": "Status",
            "type": "string"
          },
          "strategy": {
            "$ref": "#/components/schemas/StrategySpec"
          },
          "trades": {
            "items": {
              "$ref": "#/components/schemas/Trade"
            },
            "maxItems": 10000,
            "title": "Trades",
            "type": "array"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "maxItems": 32,
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "request_id",
          "job_id",
          "run_id",
          "data_version",
          "strategy",
          "data_profile",
          "metrics",
          "result_digest",
          "duration_ms"
        ],
        "title": "BacktestExecuteResponse",
        "type": "object"
      },
      "BacktestMetrics": {
        "properties": {
          "final_equity": {
            "title": "Final Equity",
            "type": "number"
          },
          "max_drawdown_pct": {
            "title": "Max Drawdown Pct",
            "type": "number"
          },
          "sharpe_ratio": {
            "title": "Sharpe Ratio",
            "type": "number"
          },
          "total_return_pct": {
            "title": "Total Return Pct",
            "type": "number"
          },
          "trade_count": {
            "title": "Trade Count",
            "type": "integer"
          },
          "win_rate_pct": {
            "title": "Win Rate Pct",
            "type": "number"
          }
        },
        "required": [
          "total_return_pct",
          "max_drawdown_pct",
          "sharpe_ratio",
          "trade_count",
          "win_rate_pct",
          "final_equity"
        ],
        "title": "BacktestMetrics",
        "type": "object"
      },
      "DataProfile": {
        "properties": {
          "duplicate_timestamps": {
            "default": 0,
            "title": "Duplicate Timestamps",
            "type": "integer"
          },
          "end_at": {
            "format": "date-time",
            "title": "End At",
            "type": "string"
          },
          "non_positive_prices": {
            "default": 0,
            "title": "Non Positive Prices",
            "type": "integer"
          },
          "row_count": {
            "title": "Row Count",
            "type": "integer"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "start_at": {
            "format": "date-time",
            "title": "Start At",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "synthetic": {
            "title": "Synthetic",
            "type": "boolean"
          },
          "timeframe": {
            "title": "Timeframe",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "source",
          "synthetic",
          "symbol",
          "timeframe",
          "row_count",
          "start_at",
          "end_at"
        ],
        "title": "DataProfile",
        "type": "object"
      },
      "EquityPoint": {
        "properties": {
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "equity": {
            "title": "Equity",
            "type": "number"
          }
        },
        "required": [
          "at",
          "equity"
        ],
        "title": "EquityPoint",
        "type": "object"
      },
      "ProblemDetails": {
        "additionalProperties": false,
        "description": "Stable RFC 7807-style error body; internal exception text is excluded.",
        "properties": {
          "code": {
            "pattern": "^[A-Z][A-Z0-9_]{2,63}$",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Detail",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 8,
                "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          },
          "status": {
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "trace_id": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{32}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          },
          "type": {
            "default": "about:blank",
            "maxLength": 256,
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "code",
          "detail"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
      "ServiceHealth": {
        "additionalProperties": false,
        "properties": {
          "dependencies": {
            "additionalProperties": {
              "enum": [
                "ok",
                "degraded",
                "unavailable"
              ],
              "type": "string"
            },
            "title": "Dependencies",
            "type": "object"
          },
          "schema_version": {
            "const": "service-health-v1",
            "default": "service-health-v1",
            "title": "Schema Version",
            "type": "string"
          },
          "service": {
            "enum": [
              "agent-service",
              "backtest-service"
            ],
            "title": "Service",
            "type": "string"
          },
          "status": {
            "enum": [
              "ok",
              "degraded"
            ],
            "title": "Status",
            "type": "string"
          },
          "version": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "service",
          "status",
          "version"
        ],
        "title": "ServiceHealth",
        "type": "object"
      },
      "StrategySpec": {
        "additionalProperties": false,
        "properties": {
          "end_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "execution": {
            "const": "next_bar_open",
            "default": "next_bar_open",
            "title": "Execution",
            "type": "string"
          },
          "fast_window": {
            "default": 20,
            "maximum": 200.0,
            "minimum": 2.0,
            "title": "Fast Window",
            "type": "integer"
          },
          "fee_bps": {
            "default": 2.0,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Fee Bps",
            "type": "number"
          },
          "initial_cash": {
            "default": 100000.0,
            "exclusiveMinimum": 0.0,
            "maximum": 100000000.0,
            "title": "Initial Cash",
            "type": "number"
          },
          "long_only": {
            "const": true,
            "default": true,
            "title": "Long Only",
            "type": "boolean"
          },
          "slippage_bps": {
            "default": 3.0,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Slippage Bps",
            "type": "number"
          },
          "slow_window": {
            "default": 60,
            "maximum": 400.0,
            "minimum": 3.0,
            "title": "Slow Window",
            "type": "integer"
          },
          "start_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "strategy_type": {
            "const": "sma_crossover",
            "default": "sma_crossover",
            "title": "Strategy Type",
            "type": "string"
          },
          "symbol": {
            "default": "XAUUSD",
            "maxLength": 32,
            "minLength": 1,
            "title": "Symbol",
            "type": "string"
          },
          "timeframe": {
            "default": "1d",
            "enum": [
              "1d",
              "1h"
            ],
            "title": "Timeframe",
            "type": "string"
          }
        },
        "title": "StrategySpec",
        "type": "object"
      },
      "Trade": {
        "properties": {
          "entry_at": {
            "format": "date-time",
            "title": "Entry At",
            "type": "string"
          },
          "entry_price": {
            "title": "Entry Price",
            "type": "number"
          },
          "exit_at": {
            "format": "date-time",
            "title": "Exit At",
            "type": "string"
          },
          "exit_price": {
            "title": "Exit Price",
            "type": "number"
          },
          "exit_reason": {
            "title": "Exit Reason",
            "type": "string"
          },
          "pnl": {
            "title": "Pnl",
            "type": "number"
          },
          "quantity": {
            "title": "Quantity",
            "type": "number"
          },
          "return_pct": {
            "title": "Return Pct",
            "type": "number"
          }
        },
        "required": [
          "entry_at",
          "exit_at",
          "entry_price",
          "exit_price",
          "quantity",
          "pnl",
          "return_pct",
          "exit_reason"
        ],
        "title": "Trade",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Internal deterministic compute service; Java owns control-plane state.",
    "title": "GoldAnalyze Backtest Engine",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health/live": {
      "get": {
        "operationId": "liveness_health_live_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceHealth"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Liveness"
      }
    },
    "/health/ready": {
      "get": {
        "operationId": "readiness_health_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceHealth"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Readiness"
      }
    },
    "/internal/v1/backtests/execute": {
      "post": {
        "operationId": "execute_backtest_internal_v1_backtests_execute_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 256,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestExecuteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "408": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request Timeout"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Execute Backtest"
      }
    }
  }
}